Split editor blocks markdown model boundaries
