+++ scripts/test-editor-blocks-model.js
@@ -3,6 +3,7 @@ import { readFileSync } from 'node:fs';
 
 import {
   applyInlineLinkToRuns,
+  autofixMarkdownSourceBlock,
   BLOCK_TYPES,
   makeBlock,
   parseInlineRuns,
@@ -10,8 +11,22 @@ import {
   serializeInlineRuns,
   serializeMarkdownBlocks
 } from '../assets/js/editor-blocks-model.js';
+import {
+  BLOCK_TYPES as coreBlockTypes,
+  makeBlock as coreMakeBlock
+} from '../assets/js/editor-blocks-block-core-model.js';
+import {
+  autofixMarkdownSourceBlock as parserAutofixMarkdownSourceBlock,
+  parseMarkdownBlocks as parserParseMarkdownBlocks
+} from '../assets/js/editor-blocks-markdown-parse-model.js';
+import {
+  serializeMarkdownBlocks as serializerSerializeMarkdownBlocks
+} from '../assets/js/editor-blocks-markdown-serialize-model.js';
 
 const source = readFileSync(new URL('../assets/js/editor-blocks-model.js', import.meta.url), 'utf8');
+const coreSource = readFileSync(new URL('../assets/js/editor-blocks-block-core-model.js', import.meta.url), 'utf8');
+const parserSource = readFileSync(new URL('../assets/js/editor-blocks-markdown-parse-model.js', import.meta.url), 'utf8');
+const serializerSource = readFileSync(new URL('../assets/js/editor-blocks-markdown-serialize-model.js', import.meta.url), 'utf8');
 
 const run = (name, fn) => {
   try {
@@ -23,7 +38,12 @@ const run = (name, fn) => {
   }
 };
 
-run('model owns block type registration and markdown round-trip rules', () => {
+run('model facade keeps block core and markdown round-trip compatibility', () => {
+  assert.equal(BLOCK_TYPES, coreBlockTypes);
+  assert.equal(makeBlock, coreMakeBlock);
+  assert.equal(parseMarkdownBlocks, parserParseMarkdownBlocks);
+  assert.equal(serializeMarkdownBlocks, serializerSerializeMarkdownBlocks);
+  assert.equal(autofixMarkdownSourceBlock, parserAutofixMarkdownSourceBlock);
   assert.equal(BLOCK_TYPES.has('paragraph'), true);
   assert.equal(BLOCK_TYPES.has('table'), true);
   assert.equal(BLOCK_TYPES.has('blank'), true);
@@ -45,14 +65,16 @@ run('model owns inline run parsing and serialization', () => {
   assert.equal(serializeInlineRuns(linked), '[Read docs](https://example.com/a%20b "Docs")');
 });
 
-run('model normalizes unknown block types without touching DOM state', () => {
+run('block core normalizes unknown block types without touching DOM state', () => {
   const block = makeBlock('unknown', 'raw', { dirty: true });
   assert.equal(block.type, 'source');
   assert.equal(block.raw, 'raw');
   assert.equal(block.dirty, true);
 });
 
-run('model stays DOM-free', () => {
-  assert.doesNotMatch(source, /\b(?:document|window|localStorage|CustomEvent)\b/);
-  assert.doesNotMatch(source, /\b(?:addEventListener|classList|querySelector|createElement|ownerDocument)\b/);
+run('model facade and extracted markdown models stay DOM-free', () => {
+  for (const modelSource of [source, coreSource, parserSource, serializerSource]) {
+    assert.doesNotMatch(modelSource, /\b(?:document|window|localStorage|CustomEvent)\b/);
+    assert.doesNotMatch(modelSource, /\b(?:addEventListener|classList|querySelector|createElement|ownerDocument)\b/);
+  }
 });
